requirement.txt
torch
colorama
skimage
cv2
tqdm
matplotlib
PyYAML
pandas
ultralytics
tensorrt-cu12